Added list of available configurations, and a test for it
